requirements.txt
python-telegram-bot==20.7
openai
python-telegram-bot==20.7
openai
gspread
google-api-python-client
google-auth
python-dotenv